fixed restart script

+++ restart-jupyter.sh
@@ -1,12 +1,15 @@
 #!/usr/bin/env bash
-for p in $(ps ux | grep '[n]otebook' | awk '{print $2}'); do kill -9 $p; done
 
 PROJECT_ROOT="$HOME/data-vis"
-
 cd $PROJECT_ROOT
 
-source venv/2.7/bin/activate && jupyter notebook 2>&1 &
-disown
+if [ ! -d $PROJECT_ROOT/venv/2.7 ]; then
+	echo "no virtualenv..."
+	exit 1
+fi
 
-deactivate
+for p in $(ps ux | grep '[n]otebook' | awk '{print $2}'); do kill -9 $p; done
 
+source venv/2.7/bin/activate && jupyter notebook --debug > notebook.log 2>&1 &disown
+
+exit 0
